refactor: move all stores to the stores folder
